added cross-pod eventual consistency

+++ server.py
@@ -5,6 +5,7 @@ import redis
 import redis.exceptions
 from redis.sentinel import Sentinel
 import os
+import time
 
 HOST = ["0.0.0.0", "127.0.0.1"][0]
 # PORT = 8080
@@ -42,6 +43,11 @@ def handle_client(client_socket, client_set, up_to_date=False):
     if not up_to_date and len(client_set) > 1:
         if canvas_state != []:
             client_socket.sendall(json.dumps(canvas_state).encode())
+    
+    if len(client_set) == 1 and json.loads(redis_client.get("client_count")) > 1: # This circumvents the stateless pods
+        if canvas_state != []:
+            client_socket.sendall(json.dumps(canvas_state).encode())
+
     up_to_date = True
     while True:
         try:
@@ -75,12 +81,26 @@ def handle_client(client_socket, client_set, up_to_date=False):
 
     print(f"removing client: {client_socket}")
     client_set.remove(client_socket)
+    redis_client.decr("client_count")
     client_socket.close()
 
-    if len(client_set) == 0:
+    if len(client_set) == 0 and json.loads(redis_client.get("client_count")) <= 0:
         print(f"No connected clients. Shutting down...")
         redis_client.flushall()
         # os._exit(0)
+
+def update_state(client_set):
+    global redis_client
+    while True:
+        time.sleep(2)
+        if len(client_set) > 0:
+            try:
+                # canvas = json.loads(redis_client.get("canvas").decode())
+                canvas = redis_client.get("canvas") # object needs to be of type bytes()
+                for client in client_set:
+                    client.sendall(canvas)
+            except:
+                print("Canvas doesn't exist")
 
 def main():
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as server_socket:
@@ -91,14 +111,21 @@ def main():
 
         client_set = set()
 
+        update_thread = threading.Thread(target=update_state, args=[client_set])
+        update_thread.daemon = True
+        update_thread.start()
+
         while True:
             try:
                 client_socket, client_address = server_socket.accept()
                 print("Accepted connection from:", client_address)
 
                 client_set.add(client_socket)
-                # if canvas_state:
-                #     client_socket.sendall(canvas_state)
+                if not redis_client.get("client_count"):
+                    redis_client.set("client_count", 1) # first client
+                else:
+                    redis_client.incr("client_count")
+
                 client_thread = threading.Thread(target=handle_client, args=(client_socket, client_set))
                 client_thread.daemon = True
                 client_thread.start()
